moved sprockets into subfolder
