+++ Extensions/Sprockets/IPityTheFoolSprocket/IPityTheFoolSprocket.csproj
