+++ Extensions/Sprockets/QuizSprocket/Properties/AssemblyInfo.cs
